Refactored simple map to map_builders mod

+++ src/spawner.rs
@@ -6,7 +6,7 @@ use specs::{Builder, World, WorldExt};
 
 pub const MAP_WIDTH: i32 = 80;
 pub const MAP_HEIGHT: i32 = 50;
-pub const MAP_COUNT: i32 = MAP_WIDTH * MAP_HEIGHT;
+pub const MAP_COUNT: usize = (MAP_WIDTH * MAP_HEIGHT) as usize;
 
 pub fn random_monster(ecs: &mut World, x: i32, y: i32) {
     let monster_name: String;
